improve load-database script

+++ scripts/load-database.js
@@ -1,26 +1,41 @@
 const spawn = require('spawn-command-with-kill')
 
-console.log('spawning')
-const mongodChild = spawn('npm start mongo --silent', {stdio: 'inherit'})
+const stopMongo = 'npm start mongo.stop'
+console.log(`stopping mongod with "${stopMongo}" (in case it is running)`)
+const stopMongoChild = spawn(stopMongo, {stdio: 'ignore'})
 
-process.on('exit', cleanUp)
-process.on('SIGINT', cleanUp)
-process.on('uncaughtException', cleanUp)
+stopMongoChild.on('exit', () => {
+  const startMongo = 'npm start mongo'
+  console.log(`starting mongod with "${startMongo}"`)
+  const mongodChild = spawn(startMongo, {stdio: 'ignore'})
+  let killed = false
 
-setTimeout(spawnGenerate, 3000)
+  process.on('exit', cleanUp)
+  process.on('SIGINT', cleanUp)
+  process.on('uncaughtException', cleanUp)
 
-function spawnGenerate() {
-  const babelNode = './node_modules/.bin/babel-node'
-  const generateChild = spawn(`cd api && ${babelNode} scripts/generate`, {
-    stdio: 'inherit',
-  })
-  generateChild.on('exit', cleanUp)
-}
+  setTimeout(spawnGenerate, 3000)
 
-function cleanUp() {
-  console.log('killing mongod process')
-  mongodChild.kill()
-  setTimeout(() => {
-    // wait for mongo to die
-  }, 1000)
-}
+  function spawnGenerate() {
+    const babelNode = './node_modules/.bin/babel-node'
+    const generateCommand = `cd api && ${babelNode} scripts/generate`
+    console.log(`generating and inserting data with "${generateCommand}"`)
+    const generateChild = spawn(generateCommand, {
+      stdio: 'inherit',
+    })
+    generateChild.on('exit', cleanUp)
+  }
+  
+  function cleanUp() {
+    if (!killed) {
+      killed = true
+      console.log('killing mongod process')
+      mongodChild.kill()
+      setTimeout(() => {
+        // wait for mongo to die
+      }, 1000)
+    }
+  }
+})
+
+
